+++ inst/www/textarea.js
@@ -4,3 +4,8 @@ const update_textarea = function(message) {
   $element.val(message.value);
 };
 Shiny.addCustomMessageHandler('update_textarea', update_textarea);
+$(document).on('shiny:inputchanged', function(event) {
+  if (event.el.classList.contains('sg_textarea')) {
+    event.preventDefault();
+  }
+});