+++ NEWS.md
@@ -1,9 +1,9 @@
 # shinyGizmo (development version)
 
 * `textArea` now stores its id as `data-id` attribute. This prevents automatic binding of the element by shiny library.
-Even though, when the id is directly specified, the input value is blocked by `preventDefault` method.
-* Add `catchIds` argument to `comminInput(s)`. The argument allows to precise which bindings should 
-be wrapped together.
+Even when the id is specified directly, the input value is blocked by `preventDefault` method.
+* Add `ignoreIds` argument to `comminInput(s)`. The argument allows to precise which bindings should 
+be ignored while merging input controllers.
 * Fix handling inherited input values with valid Shiny Input Handlers. This applies to `commonInput(s)` 
 and `valueButton`.